requirements.txt
streamlit
google-generativeai
python-dotenv
youtube_transcript_api
pathlib